Extracted common functions from bibtex parse
