+++ _includes/scripts.html
@@ -52,5 +52,5 @@
 {% endif %}
 
 {% if page.bibtex_paper %}
-<script src="{{ '/assets/js/bibtex-paper.js' | relative_url }}"></script>
+<script type="module" src="{{ '/assets/js/bibtex-paper.js' | relative_url }}"></script>
 {% endif %}
